requirements.txt
# Core dependencies for the Destination Dashboard PoC
pandas>=2.0.0
plotly>=5.14.0
requests>=2.31.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
responses>=0.24.0

# Development tools
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
pre-commit>=3.0.0
python-dotenv>=1.0.0
